refactor: BLOG-46 post page에 layout.tsx 적용 및 loading indicator 추가
