+++ package.json
@@ -14,7 +14,7 @@
     "@types/react-syntax-highlighter": "^15.5.8",
     "axios": "^1.5.1",
     "github-markdown-css": "^5.3.0",
-    "junyeol-components": "^0.0.80",
+    "junyeol-components": "^0.1.24",
     "next": "latest",
     "react": "latest",
     "react-datocms": "^4.1.3",
