+++ src/app/_components/Cards/Card/index.tsx
@@ -1,3 +1,4 @@
+"use client";
 import {
   CardProps as StorybookCardProps,
   Card as StorybookCard,
