+++ src/app/_components/LoadingCircle/index.module.scss
@@ -1,0 +1,26 @@
+@use "src/styles/libs" as *;
+
+.loading-indicator {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+
+  .spinner {
+    border: 4px solid $GRAY-200;
+    border-top: 4px solid $BLUE-400;
+    border-radius: 50%;
+    width: 40px;
+    height: 40px;
+    animation: spin 1s linear infinite;
+  }
+
+  @keyframes spin {
+    0% {
+      transform: rotate(0deg);
+    }
+    100% {
+      transform: rotate(360deg);
+    }
+  }
+}
