+++ src/app/_components/LoadingCircle/index.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import styles from "./index.module.scss";
+
+export const LoadingCircle = () => {
+  return (
+    <div className={styles["loading-indicator"]}>
+      <div className={styles["spinner"]} />
+    </div>
+  );
+};
