+++ src/app/_components/Post/index.tsx
@@ -1,8 +1,6 @@
 import React from "react";
-import styles from "./index.module.scss";
 import "github-markdown-css/github-markdown-light.css";
-
-import { Markdown } from "@/app/post/[id]/Markdown";
+import { Markdown } from "@/app/_components/Post/Markdown";
 import { PostType } from "@/types";
 
 interface PostProps {
@@ -14,11 +12,5 @@ interface PostProps {
 export default function Post({ article }: PostProps) {
   const { markdown } = article;
 
-  return (
-    <div className={`${styles.post_wrapper} markdown-body `}>
-      <section>
-        <Markdown>{markdown}</Markdown>
-      </section>
-    </div>
-  );
+  return <Markdown>{markdown}</Markdown>;
 }
