+++ src/app/post/[id]/page.module.scss
